+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "musicchartsite",
 	"description": "A website to create Music Charts",
-	"version": "0.6.0",
+	"version": "0.6.1",
 	"author": {
 		"email": "musicChart@danielTurcich.com",
 		"name": "Daniel Turcich",
@@ -40,22 +40,22 @@
 	"devDependencies": {
 		"@nabla/vite-plugin-eslint": "^1.4.1",
 		"@types/node": "^18.7.18",
-		"@typescript-eslint/eslint-plugin": "^5.37.0",
-		"@typescript-eslint/parser": "^5.37.0",
+		"@typescript-eslint/eslint-plugin": "^5.38.0",
+		"@typescript-eslint/parser": "^5.38.0",
 		"@vitejs/plugin-vue": "^3.1.0",
 		"@vue/eslint-config-typescript": "^11.0.1",
-		"autoprefixer": "^10.4.11",
+		"autoprefixer": "^10.4.12",
 		"eslint": "^8.23.1",
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-prettier": "^4.2.1",
-		"eslint-plugin-vue": "^9.4.0",
+		"eslint-plugin-vue": "^9.5.1",
 		"husky": "^8.0.1",
 		"postcss": "^8.4.16",
 		"prettier": "^2.7.1",
 		"prettier-plugin-tailwindcss": "^0.1.13",
 		"tailwindcss": "^3.1.8",
 		"typescript": "^4.8.3",
-		"vite": "^3.1.1",
+		"vite": "^3.1.3",
 		"vue-eslint-parser": "^9.1.0",
 		"vue-tsc": "^0.40.13"
 	}
