+++ src/App.vue
@@ -1,23 +1,21 @@
 <script setup lang="ts">
 import { ref, watch } from 'vue'
-import { GlobalSiteOptions } from './shared'
-import { GetSiteOptions, SaveSiteOptions } from './storage'
+import { GlobalSiteOptions } from '#shared/globals'
+import { SaveSiteOptions } from '#shared/storage'
 
 import Chart from '#components/Chart.vue'
 import Sidebar from '#components/Sidebar.vue'
 
 const renderChart = ref(false)
 
-GlobalSiteOptions.value = GetSiteOptions()
-
 // TODO: look into feasibility of debouncing this as to reduce ops
 watch(
 	GlobalSiteOptions,
 	() => {
-		if (!GlobalSiteOptions.value) {
+		if (!GlobalSiteOptions) {
 			return console.error(
 				'Error getting GlobalChartState in watch(GlobalChartState)',
-				GlobalSiteOptions.value
+				GlobalSiteOptions
 			)
 		}
