+++ src/components/Chart.vue
@@ -1,27 +1,23 @@
 <script setup lang="ts">
 import { watch } from 'vue'
-import { GlobalChartState } from '#src/shared'
-import { setStoredChart } from '#src/storage'
+import { GlobalChartState } from '#shared/globals'
+import { setStoredChart } from '#shared/storage'
 
 import ChartTitles from './ChartComponents/ChartTitles.vue'
 import ChartImages from './ChartComponents/ChartImages.vue'
 
-// TODO: create option to disable auto save maybe
 // TODO: look into feasibility of debouncing this as to reduce ops
 watch(
 	GlobalChartState,
 	() => {
-		if (!GlobalChartState.value) {
+		if (!GlobalChartState) {
 			return console.error(
 				'Error getting GlobalChartState in watch(GlobalChartState)',
 				GlobalChartState
 			)
 		}
 
-		setStoredChart(
-			GlobalChartState.value.options.chartTitle,
-			GlobalChartState.value
-		)
+		setStoredChart(GlobalChartState.options.chartTitle, GlobalChartState)
 	},
 	{
 		deep: true
