+++ src/main.ts
@@ -3,6 +3,7 @@ import App from './App.vue'
 import './tailwind.css'
 import { version as PackageVersion } from '#root/package.json'
 
+// adds a meta tag with current package.json version
 const meta = document.createElement('meta')
 meta.name = 'SiteVersion'
 meta.content = PackageVersion
